requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
sqlalchemy>=1.4.23
pydantic[email]>=1.8.2
python-dotenv>=0.19.0
jinja2>=3.0.1
python-multipart>=0.0.5
crewai>=0.1.0
langchain>=0.1.0
openai>=1.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
crewai-tools